Refactor slugify functions to common utils file a minor cleanup
